requirements/base.txt
Django==2.2
django-environ==0.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djongo==1.2.36
